requirements.txt
allure-pytest==2.9.45
pytest==6.2.5
selenium==4.1.0
requests==2.25.1
pytest-xdist==2.2.1